+++ JFinalUIBV2/platform/com/platform/plugin/ParamInitPlugin.java
@@ -127,9 +127,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(User.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_user ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_user ", splitDataSize);
 		List<User> userList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) || db_type.equals(ConstantInit.db_type_mysql)){
 				userList = User.dao.find(sql, splitDataSize, i * splitDataSize); // start 0
 				
@@ -155,9 +155,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(UserInfo.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_userInfo ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_userInfo ", splitDataSize);
 		List<UserInfo> userInfoList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) ||db_type.equals(ConstantInit.db_type_mysql)){
 				userInfoList = UserInfo.dao.find(sql, splitDataSize, i * splitDataSize); // start 0
 				
@@ -183,9 +183,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(Group.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_group ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_group ", splitDataSize);
 		List<Group> groupList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) || db_type.equals(ConstantInit.db_type_mysql)){
 				groupList = Group.dao.find(sql, splitDataSize, i * splitDataSize);
 				
@@ -211,9 +211,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(Role.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_role ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_role ", splitDataSize);
 		List<Role> roleList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) || db_type.equals(ConstantInit.db_type_mysql)){
 				roleList = Role.dao.find(sql, splitDataSize, i * splitDataSize);
 				
@@ -238,9 +238,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(Station.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_station ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_station ", splitDataSize);
 		List<Station> stationList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) || db_type.equals(ConstantInit.db_type_mysql)){
 				stationList = Station.dao.find(sql, splitDataSize, i * splitDataSize);
 				
@@ -266,9 +266,9 @@ public class ParamInitPlugin implements IPlugin {
 		String db_type = PropKit.get(ConstantInit.db_type_key);
 		String sql = ToolSqlXml.getSql(Operator.sqlId_paging, null, ConstantRender.sql_renderType_beetl);
 		
-		int batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_operator ", splitDataSize);
+		long batchCount = BaseService.service.getBatchCount(ConstantInit.db_dataSource_main, " from pt_operator ", splitDataSize);
 		List<Operator> operatorList = null;
-		for (int i = 0; i < batchCount; i++) {
+		for (long i = 0; i < batchCount; i++) {
 			if(db_type.equals(ConstantInit.db_type_postgresql) || db_type.equals(ConstantInit.db_type_mysql)){
 				operatorList = Operator.dao.find(sql, splitDataSize, i * splitDataSize);
